EIMT: Beaver: Specify which Ethernet adapter to use on old branches
L4T U-Boot is old enough that PCIe auto-probes and we end up with both a
PCIe and a USB Ethernet device. By default, U-Boot happens to pick the one
we don't want to use, so we must tell it which to use by setting
environment variable ethact for network tests. We don't define this
variable statically, since the Ethernet device naming is different in the
branches (mainline) that don't have this problem, so this value is not
valid there:-(

Signed-off-by: Stephen Warren <swarren@nvidia.com>