Add efi_loader helloworld link to travis env
We have a new test framework for efi_loader in test/py now. That framework
can fetch a hello world application via tftp, but it needs to have it
referenced in the environment.

Fortunately U-Boot builds said hello world efi binary as part of its build
process, so let's expose that to the test case.

Signed-off-by: Alexander Graf <agraf@suse.de>
Signed-off-by: Stephen Warren <swarren@nvidia.com>