push-flashair: pass request.get() params as keyword arg
According to the documentation, request.get()'s "params" parameter is a
keyword argument, not a positional parameter. Adjust the code to pass it
as such. Presumably this actually matters only in some versions of the
requests library, but not in others, since the existing code works for
some people but not others.

Reported-by: Michal Simek <michal.simek@xilinx.com>
(Michal also suggested the patch by email)
Signed-off-by: Stephen Warren <swarren@nvidia.com>