vexpress_a15: Disable HYP mode for now
QEMU v3.1.0 by default exposed EL2 (HYP) support in the A7 and A15 CPU
models.

For unknown reasons, that regresses the QEMU bootefi test case. Because
it is an unrelated bug, the best option for now is to force everything
back to the old behavior and then figure out concurrently what is broken.

Once we know we have a working base with HYP enabled, we can reenable it
for the test case again.

Signed-off-by: Alexander Graf <agraf@suse.de>